app/backen/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
django-cors-headers==3.10.1